tests - create table test refactored
